Remove delegates and fix tests.
Optional callbacks seems somehow cleaner than all the delegates so it’s implemented as such on the list view to demonstrate the difference. It also then allows easy insertion of Rx style code if desired.

Fixed up the tests that often failed once I realised the model was doing Async (to mimic things in the UI) but which screwed the tests up. Testable models used since the expectation and wait wouldn’t have worked given the Async was locked into the model code.